# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Declarative Forms API
  version: 1.0.0
  description: Render and manage declarative forms from JSON or GitHub-hosted YAML.
servers:
  - url: /api/v1
paths:
  /health:
    get:
      summary: Check API health
      responses:
        "200":
          description: Healthy
  /forms:
    get:
      summary: List API-managed forms
      security:
        - ApiKey: []
      responses:
        "200":
          description: Managed forms ordered by most recently updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagedForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create a form from a JSON definition
      security:
        - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormDefinition"
      responses:
        "201":
          description: Form created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/InvalidForm"
  /forms/github:
    post:
      summary: Register a private GitHub YAML form
      description: Uses the server-side GITHUB_TOKEN. The token is never stored with the form.
      security:
        - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [owner, repository, path]
              properties:
                owner:
                  type: string
                repository:
                  type: string
                path:
                  type: string
                  description: YAML path, with or without the .yaml suffix
                ref:
                  type: string
                  default: main
      responses:
        "201":
          description: Registered source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GitHubSource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: GitHub file unavailable
        "422":
          $ref: "#/components/responses/InvalidForm"
        "503":
          description: GITHUB_TOKEN is not configured
  /forms/{id}:
    parameters:
      - $ref: "#/components/parameters/FormId"
    get:
      summary: Get a form for rendering
      security: []
      responses:
        "200":
          description: Form definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormDefinition"
        "404":
          description: Form not found
    put:
      summary: Replace an API-managed form
      security:
        - ApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormDefinition"
      responses:
        "200":
          description: Form updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedForm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Managed form not found
        "422":
          $ref: "#/components/responses/InvalidForm"
    delete:
      summary: Delete an API-managed form
      security:
        - ApiKey: []
      responses:
        "204":
          description: Form deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Managed form not found
  /forms/{owner}/{repository}/{path}:
    parameters:
      - name: owner
        in: path
        required: true
        schema:
          type: string
      - name: repository
        in: path
        required: true
        schema:
          type: string
      - name: path
        in: path
        required: true
        description: Public GitHub YAML path, with or without the .yaml suffix
        schema:
          type: string
    get:
      summary: Load and register a public GitHub YAML form
      security: []
      responses:
        "200":
          description: Form definition with its stable a-prefixed ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormDefinition"
        "404":
          description: Public GitHub file unavailable
        "422":
          $ref: "#/components/responses/InvalidForm"
  /forms/{id}/submissions:
    parameters:
      - $ref: "#/components/parameters/FormId"
    get:
      summary: List a form's submissions
      security:
        - ApiKey: []
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Form not found
    post:
      summary: Create or update a respondent submission
      security: []
      parameters:
        - name: id
          in: query
          description: Existing submission ID for an update
          schema:
            type: string
        - name: partial
          in: query
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Saved submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "422":
          description: Submission rejected
  /forms/{id}/submissions/{submissionId}:
    parameters:
      - $ref: "#/components/parameters/FormId"
      - name: submissionId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get one submission
      security:
        - ApiKey: []
      responses:
        "200":
          description: Submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Submission not found
  /files/upload:
    post:
      summary: Upload a respondent file
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Same-origin file URL
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url:
                    type: string
  /files/{key}:
    parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Download an uploaded file
      security: []
      responses:
        "200":
          description: Uploaded file bytes
        "404":
          description: File not found
  /email-challenges/send:
    post:
      summary: Send a respondent email verification code
      security: []
      responses:
        "200":
          description: Challenge created
  /email-challenges/verify:
    post:
      summary: Verify a respondent email code
      security: []
      responses:
        "200":
          description: Email verified
components:
  securitySchemes:
    ApiKey:
      type: http
      scheme: bearer
      description: The deployment API_KEY
  parameters:
    FormId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: "^[af]"
  responses:
    Unauthorized:
      description: Missing or invalid API_KEY
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidForm:
      description: Invalid form definition
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    FormDefinition:
      type: object
      required: [sections]
      additionalProperties: true
      properties:
        id:
          type: string
          readOnly: true
        title: {}
        sections:
          type: array
          minItems: 1
          items:
            type: object
            required: [id, fields]
            additionalProperties: true
            properties:
              id:
                type: string
                minLength: 1
              fields:
                type: array
                items:
                  type: object
                  required: [id, type]
                  additionalProperties: true
                  properties:
                    id:
                      type: string
                    type:
                      type: string
    ManagedForm:
      allOf:
        - $ref: "#/components/schemas/FormDefinition"
        - type: object
          required: [id, created_at, updated_at]
          properties:
            id:
              type: string
              pattern: "^f"
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    GitHubSource:
      type: object
      required: [id, owner, repository, path, ref]
      properties:
        id:
          type: string
          pattern: "^a"
        owner:
          type: string
        repository:
          type: string
        path:
          type: string
        ref:
          type: string
    Submission:
      type: object
      required: [id, form_id, data, status, created_at, updated_at]
      properties:
        id:
          type: string
        form_id:
          type: string
        data:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [partial, completed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: string